Add userReview component
